Return early for matching tags of quality 1.0
Modern browsers send language tags in the Accept-Language header sorted
by quality value. Al can take advantage of this by returning early if
it finds a matching language tag with a quality value of 1.0 (or
higher).

This commit also removes an unnecessary next statement at the end of a
loop in the strict_pick method.